fix flows questionnaire search choking on consent only
